Renamed file to monsterInfoAPI.js
